changed application name to snake case
