+++ social_media_application/serializers/notification.py
@@ -1,5 +1,5 @@
 from .. import ma
-from SocialMediaApplication.models import User
+from social_media_application.models import User
 from marshmallow import fields
 
 class NotificationSchema(ma.Schema):
