+++ social_media_application/views/auth.py
