+++ social_media_application/views/connection.py
