+++ social_media_application/views/post_type.py
