+++ social_media_application/views/search.py
